+++ src/test/java/stringcalculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package stringcalculator;
 
 import java.util.Arrays;
 import java.util.regex.Matcher;
