+++ src/test/java/stringcalculator/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package stringcalculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
