+++ src/test/java/stringcalculator/step2.md
